src/session-manager/requirements.txt
grpcio     
grpcio-tools
grpcio-health-checking==1.19.0
SQLAlchemy
nose